Worked on VisualBERT

+++ Nish/setup.sh
@@ -1,7 +1,11 @@
 apt-get update
-apt-get install python3.10-dev libgl1-mesa-glx unzip vim --yes
-pip install transformers opencv-python --quiet
+apt-get install python3.10-dev libgl1-mesa-glx unzip python3-setuptools vim --yes
+pip install transformers opencv-python scikit-learn datasets ipdb --quiet
+python -m pip install --upgrade pip
 python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'
-wget http://cs231n.stanford.edu/tiny-imagenet-200.zip
-unzip tiny-imagenet-200.zip
-rm tiny-imagenet-200.zip
+if [ ! -d "tiny-imagenet-200" ]
+then
+    wget http://cs231n.stanford.edu/tiny-imagenet-200.zip
+    unzip tiny-imagenet-200.zip
+    rm tiny-imagenet-200.zip
+fi
